Updated submodules. Main differences:
Changed the setup such that the geo file will be meshed based on settings given in the front end (for DG that is, DE still uses lc = 1). Also added the geo_path to the json.

Also changed the mesh-creation code to include for the possibility of having an lc value in the .geo file. As this would influence the eventual mesh, this is compensated for now.

DG changes:
- Made meshsize dependent on frequency limit
- Included Wouters IR correction code

DE changes:
- Calculating recording time based on impulse response length
- Using ceil to calculate number of samples in the auralization calculation.